Move DomInfoUpdateTask class to a separate file (#552)
* Move DomInfoUpdateTask class to a separate file


* Improved code coverage

* Improved code coverage

---------
